split cli & add tests
