+++ setup.cfg
@@ -25,8 +25,6 @@ branch = true
 source = deepqmc
 omit =
     */deepqmc/wf/paulinet/backflow.py
-    */deepqmc/cli.py
-    */deepqmc/defaults.py
     */deepqmc/torchext/sloglindet.py
     */deepqmc/torchext/bdet.py
     */deepqmc/torchext/cuda.py
